+++ packages/@office-iss/react-native-win32/src-win/Libraries/Text/TextNativeComponent.win32.js
@@ -83,12 +83,6 @@ const textViewConfig = {
     textStyle: true,
     tooltip: true,
     // Windows]
-    // boxShadow is currently per-component on Android instead of being on BaseViewConfig yet
-    ...(Platform.OS === 'android' && {
-      experimental_boxShadow: {
-        process: require('../StyleSheet/processBoxShadow').default,
-      },
-    }),
   },
   directEventTypes: {
     topTextLayout: {