+++ packages/sample-custom-component/babel.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  presets: ['module:@rnw-scripts/babel-react-native-config'],
+};