+++ packages/sample-custom-component/src/DrawingIslandNativeComponent.ts
@@ -1,0 +1,7 @@
+import codegenNativeComponent from 'react-native/Libraries/Utilities/codegenNativeComponent';
+import type { ViewProps } from 'react-native';
+
+export interface DrawingIslandProps extends ViewProps {
+}
+
+export default codegenNativeComponent<DrawingIslandProps>('DrawingIsland');