+++ vnext/Microsoft.ReactNative/Fabric/AbiComponentDescriptor.cpp
@@ -162,7 +162,7 @@ facebook::react::State::Shared AbiComponentDescriptor::createState(
 facebook::react::ShadowNodeFamily::Shared AbiComponentDescriptor::createFamily(
     facebook::react::ShadowNodeFamilyFragment const &fragment) const {
   auto eventEmitter = std::make_shared<const ConcreteEventEmitter>(
-      std::make_shared<facebook::react::EventTarget>(fragment.instanceHandle), eventDispatcher_);
+      std::make_shared<facebook::react::EventTarget>(fragment.instanceHandle, fragment.surfaceId), eventDispatcher_);
   return std::make_shared<facebook::react::ShadowNodeFamily>(
       fragment, std::move(eventEmitter), eventDispatcher_, *this);
 }