+++ vnext/Microsoft.ReactNative/Fabric/Composition/CompositionViewComponentView.cpp
@@ -518,7 +518,7 @@ static winrt::com_ptr<ID2D1PathGeometry> GenerateRoundedRectPathGeometry(
 }
 
 RoundedPathParameters GenerateRoundedPathParameters(
-    const facebook::react::RectangleCorners<float> &baseRadius,
+    const facebook::react::RectangleCorners<facebook::react::CornerRadii> &baseRadius,
     const facebook::react::RectangleEdges<float> &inset,
     const facebook::react::Size &pathSize) noexcept {
   RoundedPathParameters result;
@@ -527,10 +527,10 @@ RoundedPathParameters GenerateRoundedPathParameters(
     return result;
   }
 
-  float totalTopRadius = baseRadius.topLeft + baseRadius.topRight;
-  float totalRightRadius = baseRadius.topRight + baseRadius.bottomRight;
-  float totalBottomRadius = baseRadius.bottomRight + baseRadius.bottomLeft;
-  float totalLeftRadius = baseRadius.bottomLeft + baseRadius.topLeft;
+  float totalTopRadius = baseRadius.topLeft.horizontal + baseRadius.topRight.horizontal;
+  float totalRightRadius = baseRadius.topRight.vertical + baseRadius.bottomRight.vertical;
+  float totalBottomRadius = baseRadius.bottomRight.horizontal + baseRadius.bottomLeft.horizontal;
+  float totalLeftRadius = baseRadius.bottomLeft.vertical + baseRadius.topLeft.vertical;
 
   float maxHorizontalRadius = std::max(totalTopRadius, totalBottomRadius);
   float maxVerticalRadius = std::max(totalLeftRadius, totalRightRadius);
@@ -543,21 +543,21 @@ RoundedPathParameters GenerateRoundedPathParameters(
 
   float maxScale = std::max(1.0f, std::max(scaleHoriz, scaleVert));
 
-  result.topLeftRadiusX = std::max(0.0f, baseRadius.topLeft / maxScale - inset.left);
-  result.topLeftRadiusY = std::max(0.0f, baseRadius.topLeft / maxScale - inset.top);
-  result.topRightRadiusX = std::max(0.0f, baseRadius.topRight / maxScale - inset.right);
-  result.topRightRadiusY = std::max(0.0f, baseRadius.topRight / maxScale - inset.top);
-  result.bottomRightRadiusX = std::max(0.0f, baseRadius.bottomRight / maxScale - inset.right);
-  result.bottomRightRadiusY = std::max(0.0f, baseRadius.bottomRight / maxScale - inset.bottom);
-  result.bottomLeftRadiusX = std::max(0.0f, baseRadius.bottomLeft / maxScale - inset.left);
-  result.bottomLeftRadiusY = std::max(0.0f, baseRadius.bottomLeft / maxScale - inset.bottom);
+  result.topLeftRadiusX = std::max(0.0f, baseRadius.topLeft.horizontal / maxScale - inset.left);
+  result.topLeftRadiusY = std::max(0.0f, baseRadius.topLeft.vertical / maxScale - inset.top);
+  result.topRightRadiusX = std::max(0.0f, baseRadius.topRight.horizontal / maxScale - inset.right);
+  result.topRightRadiusY = std::max(0.0f, baseRadius.topRight.vertical / maxScale - inset.top);
+  result.bottomRightRadiusX = std::max(0.0f, baseRadius.bottomRight.horizontal / maxScale - inset.right);
+  result.bottomRightRadiusY = std::max(0.0f, baseRadius.bottomRight.vertical / maxScale - inset.bottom);
+  result.bottomLeftRadiusX = std::max(0.0f, baseRadius.bottomLeft.horizontal / maxScale - inset.left);
+  result.bottomLeftRadiusY = std::max(0.0f, baseRadius.bottomLeft.vertical / maxScale - inset.bottom);
 
   return result;
 }
 
 static winrt::com_ptr<ID2D1PathGeometry> GenerateRoundedRectPathGeometry(
     winrt::Microsoft::ReactNative::Composition::Experimental::ICompositionContext &compContext,
-    const facebook::react::RectangleCorners<float> &baseRadius,
+    const facebook::react::RectangleCorners<facebook::react::CornerRadii> &baseRadius,
     const facebook::react::RectangleEdges<float> &inset,
     const facebook::react::RectangleEdges<float> &rectPathGeometry) noexcept {
   RoundedPathParameters params = GenerateRoundedPathParameters(
@@ -774,11 +774,11 @@ void DrawAllBorderLayers(
       spTextures[0], // Target Layer, Source Texture, Target Texture
       {0,
        0,
-       borderRadii.topLeft + borderWidths.left,
-       borderRadii.topLeft + borderWidths.top}, // Texture Left, Top, Width, Height
+       borderRadii.topLeft.vertical + borderWidths.left,
+       borderRadii.topLeft.horizontal + borderWidths.top}, // Texture Left, Top, Width, Height
       {AnchorPosition::Left, AnchorPosition::Top}, // Layer Anchor Point
       {0, 0}, // Layer Anchor Offset
-      {borderRadii.topLeft + borderWidths.left, borderRadii.topLeft + borderWidths.top}, // size
+      {borderRadii.topLeft.vertical + borderWidths.left, borderRadii.topLeft.horizontal + borderWidths.top}, // size
       {0.0f, 0.0f}, // relativeSize
       std::max(borderWidths.left, borderWidths.top),
       borderColors.left ? borderColors.left : borderColors.top,
@@ -791,13 +791,13 @@ void DrawAllBorderLayers(
       spBorderLayers[1],
       shape,
       spTextures[1],
-      {borderRadii.topLeft + borderWidths.left,
+      {borderRadii.topLeft.vertical + borderWidths.left,
        0,
-       textureWidth - (borderRadii.topRight + borderWidths.right),
+       textureWidth - (borderRadii.topRight.vertical + borderWidths.right),
        borderWidths.top},
       {AnchorPosition::Left, AnchorPosition::Top},
-      {borderRadii.topLeft + borderWidths.left, 0},
-      {-(borderRadii.topLeft + borderWidths.left + borderRadii.topRight + borderWidths.right),
+      {borderRadii.topLeft.vertical + borderWidths.left, 0},
+      {-(borderRadii.topLeft.vertical + borderWidths.left + borderRadii.topRight.vertical + borderWidths.right),
        borderWidths.top}, // size
       {1.0f, 0.0f}, // relativeSize
       borderWidths.top,
@@ -811,13 +811,13 @@ void DrawAllBorderLayers(
       spBorderLayers[2],
       shape,
       spTextures[2],
-      {textureWidth - (borderRadii.topRight + borderWidths.right),
+      {textureWidth - (borderRadii.topRight.vertical + borderWidths.right),
        0,
        textureWidth,
-       borderRadii.topRight + borderWidths.top},
+       borderRadii.topRight.horizontal + borderWidths.top},
       {AnchorPosition::Right, AnchorPosition::Top},
-      {-(borderRadii.topRight + borderWidths.right), 0},
-      {borderRadii.topRight + borderWidths.right, borderRadii.topRight + borderWidths.top},
+      {-(borderRadii.topRight.vertical + borderWidths.right), 0},
+      {borderRadii.topRight.vertical + borderWidths.right, borderRadii.topRight.horizontal + borderWidths.top},
       {0.0f, 0.0f},
       std::max(borderWidths.right, borderWidths.top),
       borderColors.right ? borderColors.right : borderColors.top,
@@ -831,13 +831,14 @@ void DrawAllBorderLayers(
       shape,
       spTextures[3],
       {textureWidth - borderWidths.right,
-       borderWidths.top + borderRadii.topRight,
+       borderWidths.top + borderRadii.topRight.horizontal,
        textureWidth,
-       textureHeight - (borderWidths.bottom + borderRadii.bottomRight)},
+       textureHeight - (borderWidths.bottom + borderRadii.bottomRight.horizontal)},
       {AnchorPosition::Right, AnchorPosition::Top},
-      {-borderWidths.right, borderWidths.top + borderRadii.topRight},
+      {-borderWidths.right, borderWidths.top + borderRadii.topRight.horizontal},
       {borderWidths.right,
-       -(borderWidths.top + borderRadii.topRight + borderWidths.bottom + borderRadii.bottomRight)}, // size
+       -(borderWidths.top + borderRadii.topRight.horizontal + borderWidths.bottom +
+         borderRadii.bottomRight.horizontal)}, // size
       {0.0f, 1.0f},
       borderWidths.right,
       borderColors.right,
@@ -850,13 +851,14 @@ void DrawAllBorderLayers(
       spBorderLayers[4],
       shape,
       spTextures[4],
-      {textureWidth - (borderWidths.right + borderRadii.bottomRight),
-       textureHeight - (borderWidths.bottom + borderRadii.bottomRight),
+      {textureWidth - (borderWidths.right + borderRadii.bottomRight.vertical),
+       textureHeight - (borderWidths.bottom + borderRadii.bottomRight.horizontal),
        textureWidth,
        textureHeight},
       {AnchorPosition::Right, AnchorPosition::Bottom},
-      {-(borderWidths.right + borderRadii.bottomRight), -(borderWidths.bottom + borderRadii.bottomRight)},
-      {borderWidths.right + borderRadii.bottomRight, borderWidths.bottom + borderRadii.bottomRight},
+      {-(borderWidths.right + borderRadii.bottomRight.vertical),
+       -(borderWidths.bottom + borderRadii.bottomRight.horizontal)},
+      {borderWidths.right + borderRadii.bottomRight.vertical, borderWidths.bottom + borderRadii.bottomRight.horizontal},
       {0, 0},
       std::max(borderWidths.right, borderWidths.bottom),
       borderColors.right ? borderColors.right : borderColors.bottom,
@@ -869,13 +871,13 @@ void DrawAllBorderLayers(
       spBorderLayers[5],
       shape,
       spTextures[5],
-      {borderWidths.left + borderRadii.bottomLeft,
+      {borderWidths.left + borderRadii.bottomLeft.vertical,
        textureHeight - borderWidths.bottom,
-       textureWidth - (borderWidths.right + borderRadii.bottomRight),
+       textureWidth - (borderWidths.right + borderRadii.bottomRight.vertical),
        textureHeight},
       {AnchorPosition::Left, AnchorPosition::Bottom},
-      {borderWidths.left + borderRadii.bottomLeft, -borderWidths.bottom},
-      {-(borderWidths.right + borderRadii.bottomLeft + borderWidths.left + borderRadii.bottomRight),
+      {borderWidths.left + borderRadii.bottomLeft.vertical, -borderWidths.bottom},
+      {-(borderWidths.right + borderRadii.bottomLeft.vertical + borderWidths.left + borderRadii.bottomRight.vertical),
        borderWidths.bottom},
       {1.0f, 0.0f},
       borderWidths.bottom,
@@ -890,12 +892,12 @@ void DrawAllBorderLayers(
       shape,
       spTextures[6],
       {0,
-       textureHeight - (borderWidths.bottom + borderRadii.bottomLeft),
-       borderWidths.left + borderRadii.bottomLeft,
+       textureHeight - (borderWidths.bottom + borderRadii.bottomLeft.horizontal),
+       borderWidths.left + borderRadii.bottomLeft.vertical,
        textureHeight},
       {AnchorPosition::Left, AnchorPosition::Bottom},
-      {0, -(borderWidths.bottom + borderRadii.bottomLeft)},
-      {borderWidths.left + borderRadii.bottomLeft, borderWidths.bottom + borderRadii.bottomLeft},
+      {0, -(borderWidths.bottom + borderRadii.bottomLeft.horizontal)},
+      {borderWidths.left + borderRadii.bottomLeft.vertical, borderWidths.bottom + borderRadii.bottomLeft.horizontal},
       {0, 0},
       std::max(borderWidths.left, borderWidths.bottom),
       borderColors.left ? borderColors.left : borderColors.bottom,
@@ -909,12 +911,13 @@ void DrawAllBorderLayers(
       shape,
       spTextures[7],
       {0,
-       borderWidths.top + borderRadii.topLeft,
+       borderWidths.top + borderRadii.topLeft.horizontal,
        borderWidths.left,
-       textureHeight - (borderWidths.bottom + borderRadii.bottomLeft)},
+       textureHeight - (borderWidths.bottom + borderRadii.bottomLeft.horizontal)},
       {AnchorPosition::Left, AnchorPosition::Top},
-      {0, borderWidths.top + borderRadii.topLeft},
-      {borderWidths.left, -(borderWidths.top + borderRadii.topLeft + borderWidths.bottom + borderRadii.bottomLeft)},
+      {0, borderWidths.top + borderRadii.topLeft.horizontal},
+      {borderWidths.left,
+       -(borderWidths.top + borderRadii.topLeft.horizontal + borderWidths.bottom + borderRadii.bottomLeft.horizontal)},
       {0, 1},
       borderWidths.left,
       borderColors.left,
@@ -923,7 +926,7 @@ void DrawAllBorderLayers(
 
 winrt::com_ptr<ID2D1GeometryGroup> GetGeometryForRoundedBorder(
     winrt::Microsoft::ReactNative::Composition::Experimental::ICompositionContext &compContext,
-    const facebook::react::RectangleCorners<float> &radius,
+    const facebook::react::RectangleCorners<facebook::react::CornerRadii> &radius,
     const facebook::react::RectangleEdges<float> &inset,
     const facebook::react::RectangleEdges<float> &thickness,
     const facebook::react::RectangleEdges<float> &rectPathGeometry) noexcept {
@@ -993,10 +996,17 @@ winrt::com_ptr<ID2D1GeometryGroup> GetGeometryForRoundedBorder(
 // Also apply scale factor to the radii at this point
 void pixelRoundBorderRadii(facebook::react::BorderRadii &borderRadii, float scaleFactor) noexcept {
   // Always round radii down to avoid spikey circles
-  borderRadii.topLeft = std::floor(borderRadii.topLeft * scaleFactor);
-  borderRadii.topRight = std::floor(borderRadii.topRight * scaleFactor);
-  borderRadii.bottomLeft = std::floor(borderRadii.bottomLeft * scaleFactor);
-  borderRadii.bottomRight = std::floor(borderRadii.bottomRight * scaleFactor);
+  borderRadii.topLeft = {
+      std::floor(borderRadii.topLeft.horizontal * scaleFactor), std::floor(borderRadii.topLeft.vertical * scaleFactor)};
+  borderRadii.topRight = {
+      std::floor(borderRadii.topRight.horizontal * scaleFactor),
+      std::floor(borderRadii.topRight.vertical * scaleFactor)};
+  borderRadii.bottomLeft = {
+      std::floor(borderRadii.bottomLeft.horizontal * scaleFactor),
+      std::floor(borderRadii.bottomLeft.vertical * scaleFactor)};
+  borderRadii.bottomRight = {
+      std::floor(borderRadii.bottomRight.horizontal * scaleFactor),
+      std::floor(borderRadii.bottomRight.vertical * scaleFactor)};
 }
 
 void scaleAndPixelRoundBorderWidths(
@@ -1084,8 +1094,10 @@ bool ComponentView::TryUpdateSpecialBorderLayers(
   float extentWidth = layoutMetrics.frame.size.width * layoutMetrics.pointScaleFactor;
   float extentHeight = layoutMetrics.frame.size.height * layoutMetrics.pointScaleFactor;
 
-  if (borderMetrics.borderRadii.topLeft != 0 || borderMetrics.borderRadii.topRight != 0 ||
-      borderMetrics.borderRadii.bottomLeft != 0 || borderMetrics.borderRadii.bottomRight != 0) {
+  if (borderMetrics.borderRadii.topLeft.horizontal != 0 || borderMetrics.borderRadii.topRight.horizontal != 0 ||
+      borderMetrics.borderRadii.bottomLeft.horizontal != 0 || borderMetrics.borderRadii.bottomRight.horizontal != 0 ||
+      borderMetrics.borderRadii.topLeft.vertical != 0 || borderMetrics.borderRadii.topRight.vertical != 0 ||
+      borderMetrics.borderRadii.bottomLeft.vertical != 0 || borderMetrics.borderRadii.bottomRight.vertical != 0) {
     if (borderStyle == facebook::react::BorderStyle::Dotted || borderStyle == facebook::react::BorderStyle::Dashed) {
       // Because in DirectX geometry starts at the center of the stroke, we need to deflate
       // rectangle by half the stroke width to render correctly.
@@ -1323,13 +1335,23 @@ bool ComponentView::getAcccessiblityIsReadOnly() noexcept {
   return true;
 }
 
+ToggleState ComponentView::getToggleState() noexcept {
+  return ToggleState::ToggleState_Off;
+}
+
+void ComponentView::Toggle() noexcept {
+  // no-op
+}
+
 void ComponentView::updateBorderLayoutMetrics(
     facebook::react::LayoutMetrics const &layoutMetrics,
     const facebook::react::ViewProps &viewProps) noexcept {
   auto borderMetrics = resolveAndAlignBorderMetrics(layoutMetrics, viewProps);
 
-  if (borderMetrics.borderRadii.topLeft == 0 && borderMetrics.borderRadii.topRight == 0 &&
-      borderMetrics.borderRadii.bottomLeft == 0 && borderMetrics.borderRadii.bottomRight == 0) {
+  if (borderMetrics.borderRadii.topLeft.horizontal == 0 && borderMetrics.borderRadii.topRight.horizontal == 0 &&
+      borderMetrics.borderRadii.bottomLeft.horizontal == 0 && borderMetrics.borderRadii.bottomRight.horizontal == 0 &&
+      borderMetrics.borderRadii.topLeft.vertical == 0 && borderMetrics.borderRadii.topRight.vertical == 0 &&
+      borderMetrics.borderRadii.bottomLeft.vertical == 0 && borderMetrics.borderRadii.bottomRight.vertical == 0) {
     Visual().as<::Microsoft::ReactNative::Composition::Experimental::IVisualInterop>()->SetClippingPath(nullptr);
   } else {
     winrt::com_ptr<ID2D1PathGeometry> pathGeometry = GenerateRoundedRectPathGeometry(
@@ -1510,11 +1532,11 @@ ViewComponentView::CreateInternalVisualHandler() const noexcept {
 void ViewComponentView::ensureVisual() noexcept {
   if (!m_visual) {
     if (m_createInternalVisualHandler) {
-      m_visual = m_createInternalVisualHandler();
+      m_visual = m_createInternalVisualHandler(*this);
     } else if (m_createVisualHandler) {
       m_visual =
           winrt::Microsoft::ReactNative::Composition::Experimental::MicrosoftCompositionContextHelper::CreateVisual(
-              m_createVisualHandler());
+              m_createVisualHandler(*this));
     } else {
       m_visual = createVisual();
     }