+++ vnext/Microsoft.ReactNative/Fabric/platform/react/renderer/textlayoutmanager/TextLayoutManager.h
@@ -45,12 +45,12 @@ class TextLayoutManager {
    * infrastructure.
    */
   LinesMeasurements measureLines(
-      const AttributedString &attributedString,
+      const AttributedStringBox &attributedStringBox,
       const ParagraphAttributes &paragraphAttributes,
       Size size) const;
 
   std::shared_ptr<void> getHostTextStorage(
-      const AttributedString &attributedString,
+      const AttributedStringBox &attributedStringBox,
       const ParagraphAttributes &paragraphAttributes,
       LayoutConstraints layoutConstraints) const;
 
@@ -67,7 +67,7 @@ class TextLayoutManager {
    * Calculates baseline of `attributedString` using native text rendering
    * infrastructure.
    */
-  Float baseline(AttributedString attributedString, ParagraphAttributes paragraphAttributes, Size size) const;
+  Float baseline(AttributedStringBox attributedStringBox, ParagraphAttributes paragraphAttributes, Size size) const;
 
   /*
    * Returns an opaque pointer to platform-specific TextLayoutManager.
@@ -84,9 +84,9 @@ class TextLayoutManager {
 #pragma endregion
 
  private:
-  static winrt::hstring GetTransformedText(const AttributedString &attributedString);
+  static winrt::hstring GetTransformedText(const AttributedStringBox &attributedStringBox);
   static void GetTextLayout(
-      const AttributedString &attributedString,
+      const AttributedStringBox &attributedStringBox,
       const ParagraphAttributes &paragraphAttributes,
       Size size,
       winrt::com_ptr<IDWriteTextLayout> &spTextLayout) noexcept;