+++ packages/@react-native-windows/automation-channel/CHANGELOG.md
@@ -1,8 +1,43 @@
 # Change Log - @react-native-windows/automation-channel
 
-This log was last generated on Sat, 31 Aug 2024 05:22:21 GMT and should not be manually modified.
+<!-- This log was last generated on Tue, 10 Sep 2024 05:15:18 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.12.206
+
+Tue, 10 Sep 2024 05:15:18 GMT
+
+### Patches
+
+- Automatically set EnableSourceLink for in-repo builds only (jthysell@microsoft.com)
+- Bump react-native-windows to v0.0.0-canary.866
+
+## 0.12.205
+
+Sat, 07 Sep 2024 05:14:11 GMT
+
+### Patches
+
+- Integrate 8/31 (34109996+chiaramooney@users.noreply.github.com)
+- Bump react-native-windows to v0.0.0-canary.865
+
+## 0.12.204
+
+Fri, 06 Sep 2024 05:14:02 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.864
+
+## 0.12.203
+
+Thu, 05 Sep 2024 05:13:45 GMT
+
+### Patches
+
+- Integrate 8/20 (34109996+chiaramooney@users.noreply.github.com)
+- Bump react-native-windows to v0.0.0-canary.863
 
 ## 0.12.202