+++ packages/@react-native-windows/codegen/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @react-native-windows/codegen
 
-This log was last generated on Wed, 14 Aug 2024 05:14:30 GMT and should not be manually modified.
+<!-- This log was last generated on Sat, 07 Sep 2024 05:14:11 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.0.0-canary.102
+
+Sat, 07 Sep 2024 05:14:11 GMT
+
+### Changes
+
+- Integrate 8/31 (34109996+chiaramooney@users.noreply.github.com)
+
+## 0.0.0-canary.101
+
+Fri, 06 Sep 2024 05:14:01 GMT
+
+### Changes
+
+- [Fabric] Add native component codegen (30809111+acoates-ms@users.noreply.github.com)
 
 ## 0.0.0-canary.100